Added a check for build success.

+++ generate_alpha_frinexapps.sh
@@ -34,19 +34,20 @@ else
     git pull
 
     # build the frinexapps dockerfile:
-    docker build --no-cache -f docker/frinexapps.Dockerfile -t frinexapps:alpha .
+    if docker build --no-cache -f docker/frinexapps.Dockerfile -t frinexapps:alpha . 
+    then 
+        # tag the alpha version with its own build version
+        alphaVersion=$(docker run --rm -w /ExperimentTemplate/gwt-cordova frinexapps:alpha /bin/bash -c "cat /ExperimentTemplate/gwt-cordova.version")
+        echo "taging as $alphaVersion"
+        docker tag frinexapps:alpha frinexapps:$alphaVersion
 
-    # tag the alpha version with its own build version
-    alphaVersion=$(docker run --rm -w /ExperimentTemplate/gwt-cordova frinexapps:alpha /bin/bash -c "cat /ExperimentTemplate/gwt-cordova.version")
-    echo "taging as $alphaVersion"
-    docker tag frinexapps:alpha frinexapps:$alphaVersion
-
-    # make the current XSD available by version number and alpha so that they can be used by frinex builds with frinexVersion
-    docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:alpha /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.xsd /FrinexBuildService/artifacts/\$(cat /ExperimentTemplate/gwt-cordova.version).xsd"
-    docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:alpha /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.html /FrinexBuildService/artifacts/\$(cat /ExperimentTemplate/gwt-cordova.version).html"
-    docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:alpha /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.xsd /FrinexBuildService/artifacts/alpha.xsd"
-    docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:alpha /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.html /FrinexBuildService/artifacts/alpha.html"
-    # make sure the local .m2 directory has the alpha jar files.
-    docker run --rm -v m2Directory:/maven/.m2/ -w /ExperimentTemplate frinexapps:alpha /bin/bash -c "mvn install -gs /maven/.m2/settings.xml"
-    docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /FrinexBuildService frinexbuild:latest /bin/bash -c "chown frinex:www-data /FrinexBuildService/artifacts/*.xsd; chown frinex:www-data /FrinexBuildService/artifacts/*.html;"
+        # make the current XSD available by version number and alpha so that they can be used by frinex builds with frinexVersion
+        docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:alpha /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.xsd /FrinexBuildService/artifacts/\$(cat /ExperimentTemplate/gwt-cordova.version).xsd"
+        docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:alpha /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.html /FrinexBuildService/artifacts/\$(cat /ExperimentTemplate/gwt-cordova.version).html"
+        docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:alpha /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.xsd /FrinexBuildService/artifacts/alpha.xsd"
+        docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:alpha /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.html /FrinexBuildService/artifacts/alpha.html"
+        # make sure the local .m2 directory has the alpha jar files.
+        docker run --rm -v m2Directory:/maven/.m2/ -w /ExperimentTemplate frinexapps:alpha /bin/bash -c "mvn install -gs /maven/.m2/settings.xml"
+        docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /FrinexBuildService frinexbuild:latest /bin/bash -c "chown frinex:www-data /FrinexBuildService/artifacts/*.xsd; chown frinex:www-data /FrinexBuildService/artifacts/*.html;"
+    fi;
 fi;
