+++ generate_latest_frinexapps.sh
@@ -34,19 +34,20 @@ else
     git pull
 
     # build the frinexapps dockerfile:
-    docker build --no-cache -f docker/frinexapps.Dockerfile -t frinexapps:latest .
+    if docker build --no-cache -f docker/frinexapps.Dockerfile -t frinexapps:latest .
+    then
+        # tag the latest version with its own build version
+        latestVersion=$(docker run --rm -w /ExperimentTemplate/gwt-cordova frinexapps:latest /bin/bash -c "cat /ExperimentTemplate/gwt-cordova.version")
+        echo "taging as $latestVersion"
+        docker tag frinexapps:latest frinexapps:$latestVersion
 
-    # tag the latest version with its own build version
-    latestVersion=$(docker run --rm -w /ExperimentTemplate/gwt-cordova frinexapps:latest /bin/bash -c "cat /ExperimentTemplate/gwt-cordova.version")
-    echo "taging as $latestVersion"
-    docker tag frinexapps:latest frinexapps:$latestVersion
-
-    # make the current XSD available by version number and latest so that they can be used by frinex builds with frinexVersion
-    docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:latest /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.xsd /FrinexBuildService/artifacts/\$(cat /ExperimentTemplate/gwt-cordova.version).xsd"
-    docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:latest /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.html /FrinexBuildService/artifacts/\$(cat /ExperimentTemplate/gwt-cordova.version).html"
-    docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:latest /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.xsd /FrinexBuildService/artifacts/latest.xsd"
-    docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:latest /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.html /FrinexBuildService/artifacts/latest.html"
-    # make sure the local .m2 directory has the latest jar files.
-    docker run --rm -v m2Directory:/maven/.m2/ -w /ExperimentTemplate frinexapps:latest /bin/bash -c "mvn install -gs /maven/.m2/settings.xml"
-    docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /FrinexBuildService frinexbuild:latest /bin/bash -c "chown frinex:www-data /FrinexBuildService/artifacts/*.xsd; chown frinex:www-data /FrinexBuildService/artifacts/*.html;"
+        # make the current XSD available by version number and latest so that they can be used by frinex builds with frinexVersion
+        docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:latest /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.xsd /FrinexBuildService/artifacts/\$(cat /ExperimentTemplate/gwt-cordova.version).xsd"
+        docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:latest /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.html /FrinexBuildService/artifacts/\$(cat /ExperimentTemplate/gwt-cordova.version).html"
+        docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:latest /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.xsd /FrinexBuildService/artifacts/latest.xsd"
+        docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /ExperimentTemplate/gwt-cordova frinexapps:latest /bin/bash -c "cp /ExperimentTemplate/ExperimentDesigner/src/test/resources/frinex-rest-output/frinex.html /FrinexBuildService/artifacts/latest.html"
+        # make sure the local .m2 directory has the latest jar files.
+        docker run --rm -v m2Directory:/maven/.m2/ -w /ExperimentTemplate frinexapps:latest /bin/bash -c "mvn install -gs /maven/.m2/settings.xml"
+        docker run --rm -v buildServerTarget:/FrinexBuildService/artifacts -w /FrinexBuildService frinexbuild:latest /bin/bash -c "chown frinex:www-data /FrinexBuildService/artifacts/*.xsd; chown frinex:www-data /FrinexBuildService/artifacts/*.html;"
+    fi;
 fi;
